Rework the MigrationExample file setup.
