+++ examples/ios/swift/Migration/MigrationExample.swift
@@ -20,8 +20,21 @@ import Foundation
 import RealmSwift
 
 struct MigrationExample {
+    
+    func addExampleDataToRealm(_ exampleData: (Realm) -> Void) {
+        let url = URL(for: RealmVersion.mostRecentVersion, usingTemplate: false)
+        let configuration = Realm.Configuration(fileURL: url, schemaVersion: UInt64(RealmVersion.mostRecentVersion.rawValue))
+        let realm = try! Realm(configuration: configuration)
 
-    static func execute() {
+        try! realm.write {
+            exampleData(realm)
+        }
+        
+        // Uncomment the following line to print the location of the newly created Realm.
+//        print("Realm created at: \(String(describing: configuration.fileURL!)).")
+    }
+
+    func performMigration() {
         for realmVersion in RealmVersion.allCases {
             let realmUrl = URL(for: realmVersion, usingTemplate: true)
             let schemaVersion = UInt64(RealmVersion.mostRecentVersion.rawValue)
@@ -29,5 +42,5 @@ struct MigrationExample {
             try! Realm.performMigration(for: realmConfiguration)
         }
     }
-
+    
 }
